# docs/openapi3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Konveyor Hub API",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.3.z"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/applications/{id}/manifest": {
      "get": {
        "tags": [
          "manifests"
        ],
        "summary": "Get the latest application Manifest.",
        "description": "Get the latest application Manifest.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Application ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "injected",
            "in": "query",
            "description": "Inject secret references",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "decrypted",
            "in": "query",
            "description": "Decrypt secret references",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Manifest"
                }
              }
            }
          }
        }
      }
    },
    "/applications/{id}/manifests": {
      "post": {
        "tags": [
          "manifests"
        ],
        "summary": "Create a manifest.",
        "description": "Create a manifest.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Application ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Manifest data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Manifest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Manifest"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "manifest"
      }
    },
...
    "/manifests": {
      "get": {
        "tags": [
          "manifests"
        ],
        "summary": "List all manifests.",
        "description": "List all manifests.\nfilters:\n- application.id",
        "parameters": [
          {
            "name": "injected",
            "in": "query",
            "description": "Inject secret references",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "decrypted",
            "in": "query",
            "description": "Decrypt secret references",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Manifest"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "manifests"
        ],
        "summary": "Create a manifest.",
        "description": "Create a manifest.",
        "requestBody": {
          "description": "Manifest data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Manifest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Manifest"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "manifest"
      }
    },
    "/manifests/{id}": {
      "get": {
        "tags": [
          "manifests"
        ],
        "summary": "Get a Manifest by ID.",
        "description": "Get a Manifest by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Manifest ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "injected",
            "in": "query",
            "description": "Inject secret references",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "decrypted",
            "in": "query",
            "description": "Decrypt secret references",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Manifest"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "manifests"
        ],
        "summary": "Update a manifest.",
        "description": "Update a manifest.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Manifest ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Manifest data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Manifest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "manifest"
      },
      "delete": {
        "tags": [
          "manifests"
        ],
        "summary": "Delete a manifest.",
        "description": "Delete a manifest.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Manifest ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        }
      }
    },
...
    "/migrationwaves": {
      "get": {
        "tags": [
          "migrationwaves"
        ],
        "summary": "List all migration waves.",
        "description": "List all migration waves.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.MigrationWave"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "migrationwaves"
        ],
        "summary": "Create a migration wave.",
        "description": "Create a migration wave.",
        "requestBody": {
          "description": "Migration Wave data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.MigrationWave"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.MigrationWave"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "migrationwave"
      }
    },
    "/migrationwaves/{id}": {
      "get": {
        "tags": [
          "migrationwaves"
        ],
        "summary": "Get a migration wave by ID.",
        "description": "Get a migration wave by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Migration Wave ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.MigrationWave"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "migrationwaves"
        ],
        "summary": "Update a migration wave.",
        "description": "Update a migration wave.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "MigrationWave id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "MigrationWave data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.MigrationWave"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "migrationWave"
      },
      "delete": {
        "tags": [
          "migrationwaves"
        ],
        "summary": "Delete a migration wave.",
        "description": "Delete a migration wave.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "MigrationWave id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "github_com_konveyor_tackle2-hub_api.Map": {
        "type": "object",
        "additionalProperties": {
          "type": "object"
        }
      },
      "github_com_konveyor_tackle2-hub_api.MigrationWave": {
        "required": [
          "endDate",
          "startDate"
        ],
        "type": "object",
        "properties": {
          "applications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
            }
          },
          "createTime": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "stakeholderGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
            }
          },
          "stakeholders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
            }
          },
          "startDate": {
            "type": "string"
          },
          "updateUser": {
            "type": "string"
          }
        }
      },
...
      "github_com_konveyor_tackle2-hub_api.Ref": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "github_com_konveyor_tackle2-hub_internal_api.Manifest": {
        "type": "object",
        "properties": {
          "application": {
            "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
          },
          "content": {
            "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Map"
          },
          "createTime": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "secret": {
            "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Map"
          },
          "updateUser": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}